Extract the test run code from  into a runner module to allow reuse from other crates
