+++ tests/tests.rs
@@ -2,23 +2,11 @@ use std::path::PathBuf;
 
 // Copyright (c) 2020 Ghaith Hachem and Mathias Rieder
 use inkwell::context::Context;
-use inkwell::execution_engine::{ExecutionEngine, JitFunction};
 use rusty::diagnostics::Diagnostician;
 use rusty::*;
 
-type MainFunction<T, U> = unsafe extern "C" fn(*mut T) -> U;
-
-#[allow(dead_code)]
-#[repr(C)]
-pub struct MainType {
-    a: [usize; 1000],
-}
-
-impl Default for MainType {
-    fn default() -> Self {
-        MainType { a: [0; 1000] }
-    }
-}
+//Import the helper run methods into the tests
+pub use rusty::runner::{MainType, compile, run, compile_and_run };
 
 mod correctness {
     mod arrays;
@@ -83,85 +71,3 @@ fn get_test_file(name: &str) -> String {
     data_path.display().to_string()
 }
 
-pub trait Compilable {
-    type T: SourceContainer;
-    fn containers(self) -> Vec<Self::T>;
-}
-
-impl Compilable for &str {
-    type T = SourceCode;
-    fn containers(self) -> Vec<Self::T> {
-        let code = Self::T::from(self);
-        vec![code]
-    }
-}
-
-impl Compilable for String {
-    type T = SourceCode;
-    fn containers(self) -> Vec<Self::T> {
-        let code = self.into();
-        vec![code]
-    }
-}
-
-impl<S: SourceContainer> Compilable for Vec<S> {
-    type T = S;
-    fn containers(self) -> Vec<Self::T> {
-        self
-    }
-}
-
-///
-/// Compiles and runs the given source
-/// Returns the std result as String
-/// Source must define a main function that takes no arguments and returns an int and string
-/// The int is the return value which can be verified
-/// The string will eventually be the Stdout of the function.
-///
-// pub fn compile(context: &Context, source: String) -> ExecutionEngine {
-//     let source: Vec<SourceCode> = vec![source.as_str().into()];
-//     compile_multi(context, source)
-// }
-
-// pub fn compile_and_run<T, U>(source: String, params: &mut T) -> U {
-//     compile_and_run_multi::<T, U, SourceCode>(vec![source.as_str().into()], params)
-// }
-
-pub fn run<T, U>(exec_engine: &ExecutionEngine, name: &str, params: &mut T) -> U {
-    unsafe {
-        let main: JitFunction<MainFunction<T, U>> = exec_engine.get_function(name).unwrap();
-        let main_t_ptr = &mut *params as *mut _;
-
-        main.call(main_t_ptr)
-    }
-}
-
-///
-/// Compiles and runs the given sources
-/// Returns the std result as String
-/// Sources must define a main function that takes no arguments and returns an int and string
-/// The int is the return value which can be verified
-/// The string will eventually be the Stdout of the function.
-///
-pub fn compile<T: Compilable>(context: &Context, source: T) -> ExecutionEngine {
-    let source = source.containers();
-    let code_gen = compile_module(
-        context,
-        source,
-        vec![],
-        None,
-        Diagnostician::null_diagnostician(),
-    )
-    .unwrap();
-    println!("{}", code_gen.module.print_to_string());
-    code_gen
-        .module
-        .create_jit_execution_engine(inkwell::OptimizationLevel::None)
-        .unwrap()
-}
-
-pub fn compile_and_run<T, U, S: Compilable>(source: S, params: &mut T) -> U {
-    let context: Context = Context::create();
-    let exec_engine = compile(&context, source);
-    run::<T, U>(&exec_engine, "main", params)
-}
